Supporting the explicit lambda x => x.get.foo
