+++ src/main/scala/zio/intellij/inspections/simplifications/SimplifyServiceWithInspection.scala
@@ -2,7 +2,7 @@ package zio.intellij.inspections.simplifications
 
 import org.jetbrains.plugins.scala.codeInspection.collections.{Simplification, SimplificationType}
 import org.jetbrains.plugins.scala.lang.psi.api.base.types.ScTypeElement
-import org.jetbrains.plugins.scala.lang.psi.api.expr.{ScExpression, ScGenericCall, ScMethodCall, ScReferenceExpression}
+import org.jetbrains.plugins.scala.lang.psi.api.expr._
 import org.jetbrains.plugins.scala.lang.psi.types.TypePresentationContext
 import org.jetbrains.plugins.scala.lang.refactoring.ScTypePresentationExt
 import zio.intellij.inspections._
@@ -10,7 +10,7 @@ import zio.intellij.inspections.hasMethods.`.get`
 import zio.intellij.utils._
 import zio.intellij.utils.types.ZioType
 
-class SimplifyAccessMInspection extends ZInspection(ServiceWithSimplificationType) {
+class SimplifyServiceWithInspection extends ZInspection(ServiceWithSimplificationType) {
   override protected def isAvailable(zioVersion: Version): Boolean = zioVersion >= Version.ZIO.`1.0.6`
 }
 
@@ -25,42 +25,49 @@ object ServiceWithSimplificationType extends SimplificationType {
   )(implicit
     ctx: TypePresentationContext = TypePresentationContext(expr)
   ): Option[Simplification] = {
-    val serviceTypeArg = for {
-      arg           <- accessTypeArg
-      tpe           <- arg.`type`().toOption
-      baseType      <- resolveAliases(tpe)
-      innerTypeArgs <- extractTypeArguments(baseType)
-      if innerTypeArgs.size == 1 // should be exactly one argument
-    } yield innerTypeArgs.head
+    val tpe = extractServiceTypeArgument(accessTypeArg)
 
-    serviceTypeArg.map { tpe =>
-      val fnCall = ref.getText.replace("_.get.", "_.")
-
-      replace(expr)
-        .withText(s"${zioType.name}.serviceWith[${tpe.codeText}])($fnCall)")
-        .highlightFrom(expr)
+    val refText = ref match {
+      case f: ScFunctionExpr => LambdaUtils.lambdaToUnderscore(f)
+      case _                 => ref.getText
     }
+
+    val fnCall = refText.replace("_.get.", "_.")
+
+    val replacement = replace(expr)
+      .withText(s"${zioType.name}.serviceWith${tpe.fold("")(t => s"[${t.codeText}]")}($fnCall)")
+      .highlightFrom(expr)
+
+    Some(replacement)
   }
 
   override def getSimplification(expr: ScExpression): Option[Simplification] =
     expr match {
-      // match both (_.get.foo) and (h => h.get.foo), also (_.get.foo(2))
       case ScMethodCall(accessCall, Seq(InvokedExpression(ref))) =>
         accessCall match {
           // ZIO.accessM(...)
-          case `ZIO.accessM`(zioType, _) => replacement(zioType, expr, ref)
+          case `ZIO.accessM`(zioType, _) =>
+            replacement(zioType, expr, ref)
           // ZIO.accessM[TypeAlias](...)
           case ScGenericCall(`ZIO.accessM`(zioType, _), Seq(accessTypeArg)) =>
             replacement(zioType, expr, ref, Some(accessTypeArg))
+
           case _ => None
         }
       case _ => None
     }
 
   object InvokedExpression {
+
+    /**
+     * 1. _.get.foo is a reference expression, qualified with _.get (or x => x.get)
+     * 2. _.get.foo(param) is a method call with an inner reference expr (1)
+     * 3. x => x.get.foo is a function expression with an inner reference expr (1)
+     */
     def unapply(expr: ScExpression): Option[ScExpression] = expr match {
       case ref @ ScReferenceExpression.withQualifier(`.get`(_) | lambda(_, `.get`(_))) => Some(ref)
       case mc @ ScMethodCall(InvokedExpression(_), _)                                  => Some(mc)
+      case fn @ ScFunctionExpr(_, Some(InvokedExpression(_)))                          => Some(fn)
       case _                                                                           => None
     }
   }
